# Initialize v1 and v2 storage only when it is used