--- a/uuid.go
+++ b/uuid.go
@@ -64,6 +64,7 @@ const dash byte = '-'
 // UUID v1/v2 storage.
 var (
 	storageMutex  sync.Mutex
+	storageOnce   sync.Once
 	epochFunc     func() uint64 = unixTimeFunc
 	clockSequence uint16
 	lastTime      uint64
@@ -78,12 +79,13 @@ var (
 	byteGroups = []int{8, 4, 4, 4, 12}
 )
 
-// Initialize storage
-func init() {
+func initClockSequence() {
 	buf := make([]byte, 2)
 	safeRandom(buf)
 	clockSequence = binary.BigEndian.Uint16(buf)
+}
 
+func initHardwareAddr() {
 	interfaces, err := net.Interfaces()
 	if err == nil {
 		for _, iface := range interfaces {
@@ -100,6 +102,11 @@ func init() {
 
 	// Set multicast bit as recommended in RFC 4122
 	hardwareAddr[0] |= 0x01
+}
+
+func initStorage() {
+	initClockSequence()
+	initHardwareAddr()
 }
 
 func safeRandom(dest []byte) {
@@ -300,8 +307,10 @@ func FromString(input string) (u UUID, err error) {
 }
 
 // Returns UUID v1/v2 storage state.
-// Returns epoch timestamp and clock sequence.
-func getStorage() (uint64, uint16) {
+// Returns epoch timestamp, clock sequence, and hardware address.
+func getStorage() (uint64, uint16, []byte) {
+	storageOnce.Do(initStorage)
+
 	storageMutex.Lock()
 	defer storageMutex.Unlock()
 
@@ -313,21 +322,21 @@ func getStorage() (uint64, uint16) {
 	}
 	lastTime = timeNow
 
-	return timeNow, clockSequence
+	return timeNow, clockSequence, hardwareAddr[:]
 }
 
 // NewV1 returns UUID based on current timestamp and MAC address.
 func NewV1() UUID {
 	u := UUID{}
 
-	timeNow, clockSeq := getStorage()
+	timeNow, clockSeq, hardwareAddr := getStorage()
 
 	binary.BigEndian.PutUint32(u[0:], uint32(timeNow))
 	binary.BigEndian.PutUint16(u[4:], uint16(timeNow>>32))
 	binary.BigEndian.PutUint16(u[6:], uint16(timeNow>>48))
 	binary.BigEndian.PutUint16(u[8:], clockSeq)
 
-	copy(u[10:], hardwareAddr[:])
+	copy(u[10:], hardwareAddr)
 
 	u.SetVersion(1)
 	u.SetVariant()
@@ -339,6 +348,8 @@ func NewV1() UUID {
 func NewV2(domain byte) UUID {
 	u := UUID{}
 
+	timeNow, clockSeq, hardwareAddr := getStorage()
+
 	switch domain {
 	case DomainPerson:
 		binary.BigEndian.PutUint32(u[0:], posixUID)
@@ -346,14 +357,12 @@ func NewV2(domain byte) UUID {
 		binary.BigEndian.PutUint32(u[0:], posixGID)
 	}
 
-	timeNow, clockSeq := getStorage()
-
 	binary.BigEndian.PutUint16(u[4:], uint16(timeNow>>32))
 	binary.BigEndian.PutUint16(u[6:], uint16(timeNow>>48))
 	binary.BigEndian.PutUint16(u[8:], clockSeq)
 	u[9] = domain
 
-	copy(u[10:], hardwareAddr[:])
+	copy(u[10:], hardwareAddr)
 
 	u.SetVersion(2)
 	u.SetVariant()