This costs a little bit, but it makes it possible to use this package for v3 and v4 without any available entropy. Network interfaces are only scanned when using v1 or v2. I don't know if these are interesting use cases, but here's the code!

Before:
```
BenchmarkNewV1   5000000               268 ns/op               0 B/op          0 allocs/op
BenchmarkNewV2   5000000               272 ns/op               0 B/op          0 allocs/op
BenchmarkNewV3   2000000               861 ns/op             144 B/op          4 allocs/op
BenchmarkNewV4   1000000              1841 ns/op              16 B/op          1 allocs/op
BenchmarkNewV5   2000000               916 ns/op             176 B/op          4 allocs/op
```

After:
```
BenchmarkNewV1   5000000               288 ns/op               0 B/op          0 allocs/op
BenchmarkNewV2   5000000               286 ns/op               0 B/op          0 allocs/op
BenchmarkNewV3   2000000               866 ns/op             144 B/op          4 allocs/op
BenchmarkNewV4   1000000              1853 ns/op              16 B/op          1 allocs/op
BenchmarkNewV5   2000000               916 ns/op             176 B/op          4 allocs/op
```